add parrallel-ish mcts

+++ oracle/win_rate/utils.py
@@ -2,5 +2,5 @@ import numpy as np
 
 
 def accuracy_score(y_true: np.ndarray, y_pred: np.ndarray) -> float:
-    _sum = np.sum(y_true == y_pred).item()
+    _sum = np.sum(y_true.astype(float) == y_pred.astype(float)).item()
     return _sum / len(y_true)
